+++ mutator/arithmetic/assign_invert.go
@@ -27,9 +27,13 @@ var assignInvertMutations = map[token.Token]token.Token{
 }
 
 // MutatorArithmeticAssignInvert implements a mutator to invert change assign statements.
-func MutatorArithmeticAssignInvert(_ *types.Package, _ *types.Info, node ast.Node) []mutator.Mutation {
+func MutatorArithmeticAssignInvert(_ *types.Package, info *types.Info, node ast.Node) []mutator.Mutation {
 	n, ok := node.(*ast.AssignStmt)
 	if !ok {
+		return nil
+	}
+
+	if n.Tok == token.ADD_ASSIGN && len(n.Lhs) > 0 && isStringExpr(info, n.Lhs[0]) {
 		return nil
 	}