+++ mutator/arithmetic/base.go
@@ -21,9 +21,13 @@ var arithmeticMutations = map[token.Token]token.Token{
 }
 
 // MutatorArithmeticBase implements a mutator to change base arithmetic.
-func MutatorArithmeticBase(_ *types.Package, _ *types.Info, node ast.Node) []mutator.Mutation {
+func MutatorArithmeticBase(_ *types.Package, info *types.Info, node ast.Node) []mutator.Mutation {
 	n, ok := node.(*ast.BinaryExpr)
 	if !ok {
+		return nil
+	}
+
+	if n.Op == token.ADD && isStringExpr(info, n) {
 		return nil
 	}
 
@@ -44,4 +48,21 @@ func MutatorArithmeticBase(_ *types.Package, _ *types.Info, node ast.Node) []mut
 			},
 		},
 	}
+}
+
+func isStringExpr(info *types.Info, expr ast.Expr) bool {
+	if expr == nil {
+		return false
+	}
+	if lit, ok := expr.(*ast.BasicLit); ok && lit.Kind == token.STRING {
+		return true
+	}
+	if info != nil {
+		if t := info.TypeOf(expr); t != nil {
+			if basic, ok := t.Underlying().(*types.Basic); ok && basic.Info()&types.IsString != 0 {
+				return true
+			}
+		}
+	}
+	return false
 }