+++ mutator/concurrency/goroutine_remove.go
@@ -20,6 +20,8 @@ func MutatorGoroutineRemove(_ *types.Package, _ *types.Info, node ast.Node) []mu
 	switch n := node.(type) {
 	case *ast.BlockStmt:
 		l = n.List
+	case *ast.CaseClause:
+		l = n.Body
 	case *ast.CommClause:
 		l = n.Body
 	default: