+++ mutator/concurrency/goroutine_remove_test.go
@@ -1,6 +1,9 @@
 package concurrency
 
 import (
+	"go/ast"
+	"go/parser"
+	"go/token"
 	"testing"
 
 	"github.com/quality-gates/mutago/v2/mutator"
@@ -20,4 +23,29 @@ func TestMutatorGoroutineRemove(t *testing.T) {
 		"../../testdata/concurrency/goroutine_remove.go",
 		2,
 	)
+}
+
+func TestMutatorGoroutineRemoveSwitch(t *testing.T) {
+	src := `package main
+func f(x int) {
+	switch x {
+	case 1:
+		go func(){}()
+	}
+}
+`
+	fset := token.NewFileSet()
+	file, err := parser.ParseFile(fset, "test.go", src, 0)
+	if err != nil {
+		t.Fatal(err)
+	}
+	var count int
+	ast.Inspect(file, func(n ast.Node) bool {
+		muts := MutatorGoroutineRemove(nil, nil, n)
+		count += len(muts)
+		return true
+	})
+	if count != 1 {
+		t.Fatalf("expected 1 mutation for go inside switch CaseClause, got %d", count)
+	}
 }